Re-coded serial communication
Now is the GUI the one that requests data to the Arduino. That way the
Arduino doesn't have to periodically send data. Instead, it will send
the required data when asked for it. In addition, the serial
communication is implemented in a different thread, separated from the
GUI. Data is transmitted to the GUI as QByteArray and further decoded.