Fix Arduino blocking on serial error
The GUI starts sending data right after the serial port is open.

However, the Arduino restarts itself when this happens, and it
takes around 1 second to start executing the code. Therefore
it gets blocked if it receives data while booting up.

We solve this by simply adding a delay after connecting, to
ensure that Arduino is already up and running.